refactor: switch to using atlas for migrations
- Updated Participant model to use uint for ID and EventID, and changed Placement to uint16.
- Simplified RankingResponse model by removing unnecessary GORM tags.
- Modified Semester model to use primaryKey for ID and adjusted budget fields to use float32 and uint8.
- Adjusted Session model to set primaryKey for ID and added not null constraints.
- Refined Structure and Blind models to use uint and added not null constraints where applicable.
- Changed Transaction model to use uint for ID and Amount, and adjusted Create/Update requests accordingly.
- Updated User model to set CreatedAt with default timestamp.
- Modified event handler to parse eventId as uint and adjusted service calls accordingly.
- Updated participants handler to use uint for eventId in ListParticipants.
- Adjusted structures handler to use uint for structureId.
- Updated transactions handler to use uint for transactionId.
- Refactored event service methods to accept uint instead of uint64 for eventId.
- Updated tests to reflect changes in data types and ensure compatibility.
- Removed main_integration.go and migrated to Atlas for database management.
- Added new migration script for schema improvements and updated migration process.
- Updated README for webapp with clearer instructions and structure.

Jira Ticket: UWPSC-11

+++ server/internal/server/structures_handler.go
@@ -47,7 +47,7 @@ func (s *apiServer) GetStructure(ctx *gin.Context) {
 	}
 
 	svc := services.NewStructureService(s.db)
-	structure, err := svc.GetStructure(structureId)
+	structure, err := svc.GetStructure(uint(structureId))
 	if err != nil {
 		ctx.AbortWithStatusJSON(err.(e.APIErrorResponse).Code, err)
 		return
@@ -69,7 +69,7 @@ func (s *apiServer) UpdateStructure(ctx *gin.Context) {
 		ctx.AbortWithStatusJSON(http.StatusBadRequest, e.InvalidRequest(err.Error()))
 		return
 	}
-	req.ID = structureId
+	req.ID = uint(structureId)
 
 	svc := services.NewStructureService(s.db)
 	structure, err := svc.UpdateStructure(&req)
